Rename credential configuring commands and keep them separate from provider commands doc: https://microsoft.sharepoint.com/:w:/t/radiuscoreteam/EVAuQrRK6tRIqiOZmjnyxjoBUfaa2jF2uiV-jhibg5qB5A?e=OfvGWa issue: NA
