+++ pkg/cli/cmd/credential/common/validation.go
