+++ pkg/cli/cmd/credential/common/validation_test.go
