+++ pkg/cli/cmd/credential/unregister/unregister.go
@@ -3,35 +3,35 @@
 // Licensed under the MIT License.
 // ------------------------------------------------------------
 
-package show
+package unregister
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/project-radius/radius/pkg/cli"
-	"github.com/project-radius/radius/pkg/cli/clients"
 	"github.com/project-radius/radius/pkg/cli/cmd/commonflags"
-	"github.com/project-radius/radius/pkg/cli/cmd/provider/common"
+	"github.com/project-radius/radius/pkg/cli/cmd/credential/common"
 	"github.com/project-radius/radius/pkg/cli/connections"
 	"github.com/project-radius/radius/pkg/cli/framework"
-	"github.com/project-radius/radius/pkg/cli/objectformats"
 	"github.com/project-radius/radius/pkg/cli/output"
 	"github.com/project-radius/radius/pkg/cli/workspaces"
 	"github.com/spf13/cobra"
 )
 
-// NewCommand creates an instance of the command and runner for the `rad provider show` command.
+// NewCommand creates an instance of the command and runner for the `rad provider delete` command.
 func NewCommand(factory framework.Factory) (*cobra.Command, framework.Runner) {
 	runner := NewRunner(factory)
 
 	cmd := &cobra.Command{
-		Use:   "show [name]",
-		Short: "Show details of a configured cloud provider",
-		Long:  "Show details of a configured cloud provider." + common.LongDescriptionBlurb,
+		Use:   "unregister",
+		Short: "Deletes a configured cloud provider credential from the Radius installation",
+		Long:  "Deletes a configured cloud provider credential from the Radius installation." + common.LongDescriptionBlurb,
 		Example: `
-# Show cloud providers details for Azure
-rad provider show azure
+# Delete Azure cloud provider credential
+rad credential unregister azure
+
+# Delete AWS cloud provider credential
+rad credential unregister aws
 `,
 		Args: cobra.ExactArgs(1),
 		RunE: framework.RunCommand(runner),
@@ -43,7 +43,7 @@ rad provider show azure
 	return cmd, runner
 }
 
-// Runner is the runner implementation for the `rad provider show` command.
+// Runner is the runner implementation for the `rad provider delete` command.
 type Runner struct {
 	ConfigHolder      *framework.ConfigHolder
 	ConnectionFactory connections.Factory
@@ -53,7 +53,7 @@ type Runner struct {
 	Kind              string
 }
 
-// NewRunner creates a new instance of the `rad provider show` runner.
+// NewRunner creates a new instance of the `rad provider delete` runner.
 func NewRunner(factory framework.Factory) *Runner {
 	return &Runner{
 		ConfigHolder:      factory.GetConfigHolder(),
@@ -62,8 +62,9 @@ func NewRunner(factory framework.Factory) *Runner {
 	}
 }
 
-// Validate runs validation for the `rad provider show` command.
+// Validate runs validation for the `rad provider delete` command.
 func (r *Runner) Validate(cmd *cobra.Command, args []string) error {
+	// Validate command line args and
 	workspace, err := cli.RequireWorkspace(cmd, r.ConfigHolder.Config, r.ConfigHolder.DirectoryConfig)
 	if err != nil {
 		return err
@@ -90,24 +91,23 @@ func (r *Runner) Validate(cmd *cobra.Command, args []string) error {
 	return nil
 }
 
-// Run runs the `rad provider show` command.
+// Run runs the `rad provider delete` command.
 func (r *Runner) Run(ctx context.Context) error {
-	r.Output.LogInfo("Showing cloud provider %q for Radius installation %q...", r.Kind, r.Workspace.FmtConnection())
+	r.Output.LogInfo("Deleting credential for cloud provider %q for Radius installation %q...", r.Kind, r.Workspace.FmtConnection())
 	client, err := r.ConnectionFactory.CreateCloudProviderManagementClient(ctx, *r.Workspace)
 	if err != nil {
 		return err
 	}
 
-	providers, err := client.Get(ctx, r.Kind)
-	if clients.Is404Error(err) {
-		return &cli.FriendlyError{Message: fmt.Sprintf("Cloud provider %q could not be found.", r.Kind)}
-	} else if err != nil {
+	deleted, err := client.Delete(ctx, r.Kind)
+	if err != nil {
 		return err
 	}
 
-	err = r.Output.WriteFormatted(r.Format, providers, objectformats.GetCloudProviderTableFormat())
-	if err != nil {
-		return err
+	if deleted {
+		r.Output.LogInfo("Cloud provider deleted.")
+	} else {
+		r.Output.LogInfo("Cloud provider %q was not found or has been already deleted.", r.Kind)
 	}
 
 	return nil
