+++ pkg/cli/cmd/credential/unregister/unregister_test.go
@@ -3,7 +3,7 @@
 // Licensed under the MIT License.
 // ------------------------------------------------------------
 
-package show
+package unregister
 
 import (
 	"context"
@@ -13,7 +13,6 @@ import (
 	"github.com/project-radius/radius/pkg/cli/clients"
 	"github.com/project-radius/radius/pkg/cli/connections"
 	"github.com/project-radius/radius/pkg/cli/framework"
-	"github.com/project-radius/radius/pkg/cli/objectformats"
 	"github.com/project-radius/radius/pkg/cli/output"
 	"github.com/project-radius/radius/pkg/cli/workspaces"
 	"github.com/project-radius/radius/test/radcli"
@@ -28,7 +27,7 @@ func Test_Validate(t *testing.T) {
 	configWithWorkspace := radcli.LoadConfigWithWorkspace(t)
 	testcases := []radcli.ValidateInput{
 		{
-			Name:          "Valid Show Command",
+			Name:          "Valid Delete Command",
 			Input:         []string{"azure"},
 			ExpectedValid: true,
 			ConfigHolder: framework.ConfigHolder{
@@ -37,7 +36,7 @@ func Test_Validate(t *testing.T) {
 			},
 		},
 		{
-			Name:          "Show Command with fallback workspace",
+			Name:          "Delete Command with fallback workspace",
 			Input:         []string{"Azure"},
 			ExpectedValid: false,
 			ConfigHolder: framework.ConfigHolder{
@@ -46,7 +45,7 @@ func Test_Validate(t *testing.T) {
 			},
 		},
 		{
-			Name:          "Show Command with unsupported provider type",
+			Name:          "Delete Command with unsupported provider type",
 			Input:         []string{"invalidProviderType"},
 			ExpectedValid: false,
 			ConfigHolder: framework.ConfigHolder{
@@ -55,7 +54,7 @@ func Test_Validate(t *testing.T) {
 			},
 		},
 		{
-			Name:          "Show Command with insufficient args",
+			Name:          "Delete Command with insufficient args",
 			Input:         []string{},
 			ExpectedValid: false,
 			ConfigHolder: framework.ConfigHolder{
@@ -64,7 +63,7 @@ func Test_Validate(t *testing.T) {
 			},
 		},
 		{
-			Name:          "Show Command with too many args",
+			Name:          "Delete Command with too many args",
 			Input:         []string{"azure", "a", "b"},
 			ExpectedValid: false,
 			ConfigHolder: framework.ConfigHolder{
@@ -82,19 +81,14 @@ func Test_Run(t *testing.T) {
 		"context": "my-context",
 	}
 
-	t.Run("Show azure provider", func(t *testing.T) {
+	t.Run("Delete azure provider", func(t *testing.T) {
 		t.Run("Exists", func(t *testing.T) {
 			ctrl := gomock.NewController(t)
 
-			provider := clients.CloudProviderResource{
-				Name:    "azure",
-				Enabled: true,
-			}
-
 			client := clients.NewMockCloudProviderManagementClient(ctrl)
 			client.EXPECT().
-				Get(gomock.Any(), "azure").
-				Return(provider, nil).
+				Delete(gomock.Any(), "azure").
+				Return(true, nil).
 				Times(1)
 
 			outputSink := &output.MockOutput{}
@@ -112,13 +106,11 @@ func Test_Run(t *testing.T) {
 
 			expected := []any{
 				output.LogOutput{
-					Format: "Showing cloud provider %q for Radius installation %q...",
+					Format: "Deleting credential for cloud provider %q for Radius installation %q...",
 					Params: []any{"azure", "Kubernetes (context=my-context)"},
 				},
-				output.FormattedOutput{
-					Format:  "table",
-					Obj:     provider,
-					Options: objectformats.GetCloudProviderTableFormat(),
+				output.LogOutput{
+					Format: "Cloud provider deleted.",
 				},
 			}
 			require.Equal(t, expected, outputSink.Writes)
@@ -128,8 +120,8 @@ func Test_Run(t *testing.T) {
 
 			client := clients.NewMockCloudProviderManagementClient(ctrl)
 			client.EXPECT().
-				Get(gomock.Any(), "azure").
-				Return(clients.CloudProviderResource{}, radcli.Create404Error()).
+				Delete(gomock.Any(), "azure").
+				Return(false, nil).
 				Times(1)
 
 			outputSink := &output.MockOutput{}
@@ -143,8 +135,19 @@ func Test_Run(t *testing.T) {
 			}
 
 			err := runner.Run(context.Background())
-			require.Error(t, err)
-			require.Equal(t, "Cloud provider \"azure\" could not be found.", err.Error())
+			require.NoError(t, err)
+
+			expected := []any{
+				output.LogOutput{
+					Format: "Deleting credential for cloud provider %q for Radius installation %q...",
+					Params: []any{"azure", "Kubernetes (context=my-context)"},
+				},
+				output.LogOutput{
+					Format: "Cloud provider %q was not found or has been already deleted.",
+					Params: []any{"azure"},
+				},
+			}
+			require.Equal(t, expected, outputSink.Writes)
 		})
 	})
 }
