--- pkg/cli/cmd/provider/delete/delete_test.go
@@ -1,153 +1,0 @@
-// ------------------------------------------------------------
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-// ------------------------------------------------------------
-
-package delete
-
-import (
-	"context"
-	"testing"
-
-	"github.com/golang/mock/gomock"
-	"github.com/project-radius/radius/pkg/cli/clients"
-	"github.com/project-radius/radius/pkg/cli/connections"
-	"github.com/project-radius/radius/pkg/cli/framework"
-	"github.com/project-radius/radius/pkg/cli/output"
-	"github.com/project-radius/radius/pkg/cli/workspaces"
-	"github.com/project-radius/radius/test/radcli"
-	"github.com/stretchr/testify/require"
-)
-
-func Test_CommandValidation(t *testing.T) {
-	radcli.SharedCommandValidation(t, NewCommand)
-}
-
-func Test_Validate(t *testing.T) {
-	configWithWorkspace := radcli.LoadConfigWithWorkspace(t)
-	testcases := []radcli.ValidateInput{
-		{
-			Name:          "Valid Delete Command",
-			Input:         []string{"azure"},
-			ExpectedValid: true,
-			ConfigHolder: framework.ConfigHolder{
-				ConfigFilePath: "",
-				Config:         configWithWorkspace,
-			},
-		},
-		{
-			Name:          "Delete Command with fallback workspace",
-			Input:         []string{"Azure"},
-			ExpectedValid: false,
-			ConfigHolder: framework.ConfigHolder{
-				ConfigFilePath: "",
-				Config:         radcli.LoadEmptyConfig(t),
-			},
-		},
-		{
-			Name:          "Delete Command with unsupported provider type",
-			Input:         []string{"invalidProviderType"},
-			ExpectedValid: false,
-			ConfigHolder: framework.ConfigHolder{
-				ConfigFilePath: "",
-				Config:         configWithWorkspace,
-			},
-		},
-		{
-			Name:          "Delete Command with insufficient args",
-			Input:         []string{},
-			ExpectedValid: false,
-			ConfigHolder: framework.ConfigHolder{
-				ConfigFilePath: "",
-				Config:         configWithWorkspace,
-			},
-		},
-		{
-			Name:          "Delete Command with too many args",
-			Input:         []string{"azure", "a", "b"},
-			ExpectedValid: false,
-			ConfigHolder: framework.ConfigHolder{
-				ConfigFilePath: "",
-				Config:         configWithWorkspace,
-			},
-		},
-	}
-	radcli.SharedValidateValidation(t, NewCommand, testcases)
-}
-
-func Test_Run(t *testing.T) {
-	connection := map[string]any{
-		"kind":    workspaces.KindKubernetes,
-		"context": "my-context",
-	}
-
-	t.Run("Delete azure provider", func(t *testing.T) {
-		t.Run("Exists", func(t *testing.T) {
-			ctrl := gomock.NewController(t)
-
-			client := clients.NewMockCloudProviderManagementClient(ctrl)
-			client.EXPECT().
-				Delete(gomock.Any(), "azure").
-				Return(true, nil).
-				Times(1)
-
-			outputSink := &output.MockOutput{}
-
-			runner := &Runner{
-				ConnectionFactory: &connections.MockFactory{CloudProviderManagementClient: client},
-				Output:            outputSink,
-				Workspace:         &workspaces.Workspace{Connection: connection},
-				Kind:              "azure",
-				Format:            "table",
-			}
-
-			err := runner.Run(context.Background())
-			require.NoError(t, err)
-
-			expected := []any{
-				output.LogOutput{
-					Format: "Deleting cloud provider %q for Radius installation %q...",
-					Params: []any{"azure", "Kubernetes (context=my-context)"},
-				},
-				output.LogOutput{
-					Format: "Cloud provider deleted.",
-				},
-			}
-			require.Equal(t, expected, outputSink.Writes)
-		})
-		t.Run("Not Found", func(t *testing.T) {
-			ctrl := gomock.NewController(t)
-
-			client := clients.NewMockCloudProviderManagementClient(ctrl)
-			client.EXPECT().
-				Delete(gomock.Any(), "azure").
-				Return(false, nil).
-				Times(1)
-
-			outputSink := &output.MockOutput{}
-
-			runner := &Runner{
-				ConnectionFactory: &connections.MockFactory{CloudProviderManagementClient: client},
-				Output:            outputSink,
-				Workspace:         &workspaces.Workspace{Connection: connection},
-				Kind:              "azure",
-				Format:            "table",
-			}
-
-			err := runner.Run(context.Background())
-			require.NoError(t, err)
-
-			expected := []any{
-				output.LogOutput{
-					Format: "Deleting cloud provider %q for Radius installation %q...",
-					Params: []any{"azure", "Kubernetes (context=my-context)"},
-				},
-				output.LogOutput{
-					Format: "Cloud provider %q was not found or has been already deleted.",
-					Params: []any{"azure"},
-				},
-			}
-			require.Equal(t, expected, outputSink.Writes)
-		})
-	})
-}
